23.06.26

+++ client/src/pages/Detail/model/deleteItem.tsx
@@ -1,20 +1,26 @@
 import axios from 'axios';
 import { LISTCRUD_URL } from '../../../common/utils/constants';
+import useDecryptToken from '../../../common/utils/customHook/useDecryptToken';
 
 export const deleteItem = async (questionId: number) => {
   try {
-    const accessToken = localStorage.getItem('accessToken');
+    const decrypt = useDecryptToken();
 
-    const headers = {
-      'ngrok-skip-browser-warning': 'true',
-      Authorization: `Bearer ${accessToken}`,
-    };
+    const encryptedToken = localStorage.getItem('accessToken');
 
-    await axios.delete(`${LISTCRUD_URL}/questions/${questionId}`, {
-      headers,
-    });
+    if (encryptedToken) {
+      const accessToken = decrypt(encryptedToken);
 
-    console.log('삭제 요청이 성공적으로 완료되었습니다.');
+      const headers = {
+        Authorization: `Bearer ${accessToken}`,
+      };
+
+      await axios.delete(`${LISTCRUD_URL}/questions/${questionId}`, {
+        headers,
+      });
+
+      console.log('삭제 요청이 성공적으로 완료되었습니다.');
+    }
   } catch (error) {
     console.log('삭제 요청을 처리하는 중 오류가 발생했습니다');
   }
